Move run() method to kernel
