+++ radix-engine/src/engine/kernel.rs
@@ -217,6 +217,193 @@ where
             }
         }
     }
+
+    pub fn run(
+        &mut self,
+        execution_entity: ExecutionEntity,
+        fn_ident: &str,
+        input: ScryptoValue,
+    ) -> Result<ScryptoValue, RuntimeError> {
+        let output = {
+            let rtn = match execution_entity {
+                ExecutionEntity::Function(type_name) => match type_name {
+                    TypeName::TransactionProcessor => TransactionProcessor::static_main(
+                        fn_ident, input, self,
+                    )
+                    .map_err(|e| match e {
+                        TransactionProcessorError::InvalidRequestData(_) => panic!("Illegal state"),
+                        TransactionProcessorError::InvalidMethod => panic!("Illegal state"),
+                        TransactionProcessorError::RuntimeError(e) => e,
+                    }),
+                    TypeName::Package => ValidatedPackage::static_main(fn_ident, input, self)
+                        .map_err(RuntimeError::PackageError),
+                    TypeName::ResourceManager => {
+                        ResourceManager::static_main(fn_ident, input, self)
+                            .map_err(RuntimeError::ResourceManagerError)
+                    }
+                    TypeName::Blueprint(package_address, blueprint_name) => {
+                        let output = {
+                            let package = self
+                                .track()
+                                .read_substate(SubstateId::Package(package_address))
+                                .package()
+                                .clone(); // TODO: remove copy
+                            let wasm_metering_params = self.fee_table().wasm_metering_params();
+                            let instrumented_code = self
+                                .wasm_instrumenter()
+                                .instrument(package.code(), &wasm_metering_params)
+                                .to_vec(); // TODO: remove copy
+                            let mut instance = self.wasm_engine().instantiate(&instrumented_code);
+                            let blueprint_abi = package
+                                .blueprint_abi(&blueprint_name)
+                                .expect("Blueprint should exist");
+                            let export_name = &blueprint_abi
+                                .get_fn_abi(fn_ident)
+                                .unwrap()
+                                .export_name
+                                .to_string();
+                            let mut runtime: Box<dyn WasmRuntime> =
+                                Box::new(RadixEngineWasmRuntime::new(
+                                    ScryptoActor::blueprint(
+                                        package_address,
+                                        blueprint_name.clone(),
+                                    ),
+                                    self,
+                                ));
+                            instance
+                                .invoke_export(&export_name, &input, &mut runtime)
+                                .map_err(|e| match e {
+                                    // Flatten error code for more readable transaction receipt
+                                    InvokeError::RuntimeError(e) => e,
+                                    e @ _ => RuntimeError::InvokeError(e.into()),
+                                })?
+                        };
+
+                        let package = self
+                            .track()
+                            .read_substate(SubstateId::Package(package_address))
+                            .package();
+                        let blueprint_abi = package
+                            .blueprint_abi(&blueprint_name)
+                            .expect("Blueprint should exist");
+                        let fn_abi = blueprint_abi.get_fn_abi(fn_ident).unwrap();
+                        if !fn_abi.output.matches(&output.dom) {
+                            Err(RuntimeError::InvalidFnOutput {
+                                fn_ident: fn_ident.to_string(),
+                                output: output.dom,
+                            })
+                        } else {
+                            Ok(output)
+                        }
+                    }
+                },
+                ExecutionEntity::Method(_, state) => match state {
+                    ExecutionState::Consumed(node_id) => match node_id {
+                        RENodeId::Bucket(..) => {
+                            Bucket::consuming_main(node_id, fn_ident, input, self)
+                                .map_err(RuntimeError::BucketError)
+                        }
+                        RENodeId::Proof(..) => Proof::main_consume(node_id, fn_ident, input, self)
+                            .map_err(RuntimeError::ProofError),
+                        _ => panic!("Unexpected"),
+                    },
+                    ExecutionState::RENodeRef(node_id) => match node_id {
+                        RENodeId::Bucket(bucket_id) => {
+                            Bucket::main(bucket_id, fn_ident, input, self)
+                                .map_err(RuntimeError::BucketError)
+                        }
+                        RENodeId::Proof(proof_id) => Proof::main(proof_id, fn_ident, input, self)
+                            .map_err(RuntimeError::ProofError),
+                        RENodeId::Worktop => {
+                            Worktop::main(fn_ident, input, self).map_err(RuntimeError::WorktopError)
+                        }
+                        RENodeId::Vault(vault_id) => Vault::main(vault_id, fn_ident, input, self)
+                            .map_err(RuntimeError::VaultError),
+                        RENodeId::Component(component_address) => {
+                            Component::main(component_address, fn_ident, input, self)
+                                .map_err(RuntimeError::ComponentError)
+                        }
+                        RENodeId::ResourceManager(resource_address) => {
+                            ResourceManager::main(resource_address, fn_ident, input, self)
+                                .map_err(RuntimeError::ResourceManagerError)
+                        }
+                        RENodeId::System => {
+                            System::main(fn_ident, input, self).map_err(RuntimeError::SystemError)
+                        }
+                        _ => panic!("Unexpected"),
+                    },
+                    ExecutionState::Component(
+                        package_address,
+                        blueprint_name,
+                        component_address,
+                    ) => {
+                        let output = {
+                            let package = self
+                                .track()
+                                .read_substate(SubstateId::Package(package_address))
+                                .package()
+                                .clone(); // TODO: remove copy
+                            let wasm_metering_params = self.fee_table().wasm_metering_params();
+                            let instrumented_code = self
+                                .wasm_instrumenter()
+                                .instrument(package.code(), &wasm_metering_params)
+                                .to_vec(); // TODO: remove copy
+                            let mut instance = self.wasm_engine().instantiate(&instrumented_code);
+                            let blueprint_abi = package
+                                .blueprint_abi(&blueprint_name)
+                                .expect("Blueprint should exist");
+                            let export_name = &blueprint_abi
+                                .get_fn_abi(fn_ident)
+                                .unwrap()
+                                .export_name
+                                .to_string();
+                            let mut runtime: Box<dyn WasmRuntime> =
+                                Box::new(RadixEngineWasmRuntime::new(
+                                    ScryptoActor::Component(
+                                        component_address,
+                                        package_address.clone(),
+                                        blueprint_name.clone(),
+                                    ),
+                                    self,
+                                ));
+                            instance
+                                .invoke_export(&export_name, &input, &mut runtime)
+                                .map_err(|e| match e {
+                                    // Flatten error code for more readable transaction receipt
+                                    InvokeError::RuntimeError(e) => e,
+                                    e @ _ => RuntimeError::InvokeError(e.into()),
+                                })?
+                        };
+
+                        let package = self
+                            .track()
+                            .read_substate(SubstateId::Package(package_address))
+                            .package();
+                        let blueprint_abi = package
+                            .blueprint_abi(&blueprint_name)
+                            .expect("Blueprint should exist");
+                        let fn_abi = blueprint_abi.get_fn_abi(fn_ident).unwrap();
+                        if !fn_abi.output.matches(&output.dom) {
+                            Err(RuntimeError::InvalidFnOutput {
+                                fn_ident: fn_ident.to_string(),
+                                output: output.dom,
+                            })
+                        } else {
+                            Ok(output)
+                        }
+                    }
+                    ExecutionState::AuthZone(frame_id) => {
+                        AuthZone::main(frame_id, fn_ident, input, self)
+                            .map_err(RuntimeError::AuthZoneError)
+                    }
+                },
+            }?;
+
+            rtn
+        };
+
+        Ok(output)
+    }
 }
 
 impl<'g, 's, W, I, C> SystemApi<'s, W, I, C> for Kernel<'g, 's, W, I, C>
@@ -391,7 +578,7 @@ where
                 self,
             );
             self.call_frames.push(frame);
-            CallFrame::run(ExecutionEntity::Function(type_name), &fn_ident, input, self)?
+            self.run(ExecutionEntity::Function(type_name), &fn_ident, input)?
         };
 
         // Process return data
@@ -906,11 +1093,10 @@ where
                 self,
             );
             self.call_frames.push(frame);
-            CallFrame::run(
+            self.run(
                 ExecutionEntity::Method(receiver, execution_state),
                 &fn_ident,
                 input,
-                self,
             )?
         };
